+++ Widget/Nextcloud/NextcloudData.swift
@@ -206,7 +206,7 @@ func getNextcloudDataEntry(isPreview: Bool, displaySize: CGSize, completion: @es
             guard let url = URL(string: urlString) else { continue }
             // Build Recent Data
             var imageRecent = UIImage()
-            if let image = NCUtilityGUI().createFilePreviewImage(ocId: file.ocId, etag: file.etag, fileNameView: file.fileName, classFile: file.classFile, status: 0, createPreviewMedia: false) {
+            if let image = NCUtility.shared.createFilePreviewImage(ocId: file.ocId, etag: file.etag, fileNameView: file.fileName, classFile: file.classFile, status: 0, createPreviewMedia: false) {
                 imageRecent = image
             } else if !file.iconName.isEmpty {
                 imageRecent = UIImage(named: file.iconName)!
